requirements.txt
flask
requests
pywebview[edgechromium]